Созданы компоненты boardPage, startPage, finishPage. Функционал из Game.jsx перенесен в startPage.jsx
