+++ src/routes/game/game.module.css
@@ -1,11 +1,1 @@
-.game {
-  display: grid;
-  place-items: center;
-}
 
-.flex {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
-}
