+++ src/routes/game/routes/finishPage/finishPage.jsx
@@ -1,0 +1,11 @@
+import s from './finishPage.module.css';
+
+function FinishPage() {
+    return (
+        <div className={s.root}>
+            <h1>Finish Game</h1>
+        </div>
+    )
+}
+
+export default FinishPage;
