+++ src/routes/game/routes/index.js
@@ -1,0 +1,3 @@
+export { default as StartPage } from './startPage/startPage';
+export { default as BoardPage } from './boardPage/boardPage';
+export { default as FinishPage } from './finishPage/finishPage';
