+++ src/routes/game/routes/startPage/startPage.module.css
@@ -1,0 +1,11 @@
+.game {
+  display: grid;
+  place-items: center;
+}
+
+.flex {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+}
